usb: dwc3: dwc3-qcom: Make dwc3-qcom a role-switch signal recipient
When switching role from host to peripheral or peripheral to host we need
to set SoC UTMI signal logic in software since this is not done
automatically by the PHY or DWC3 core.

We have existing extcon code in dwc3-qcom which already implements the
right logic for extcon based systems, however, as we move to USB
role-switching we need to similarly facilitate the same UTMI switch
notification.

Setting the dwc3-qcom wrapper up as a USB role switch signal recipient
allows us to replicate the extcon logic with the role-switch API by
receiving the set_role() from dwc3-core and calling the existing VBUS
extcon code.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>