ASoC: codecs: wcd934x: Use .prepare and .shutdown instead of .trigger
.trigger only gets called when the actual PCM stream gets read/written.
The WIP q6voice driver is kind of a virtual FE for the modem and userspace
doesn't actually read/write PCM streams into it. So the .trigger doesn't get
called and mic doesnt work on calls. We currently use a userspace daemon called
q6voiced to open and prepare the q6voice virtual FE which in turn open and
prepares the BE speaker and mic. So for now, contain the enabling and disabling
of wcd934x as part of .prepare and .shutdown instead of .trigger.