sound: soc: codecs: wcd-mbhc-v2: expose ALSA control for jack
Jack detection is currently fully functional via the input
interface together with the multimedia buttons, but is not exposed
as an ALSA control. Therefore, ALSA clients such as PulseAudio
do not pick up the jack detection events as they only support
one of the possible interface (ALSA control or input interface,
but not both). Expose the jack events as a new ALSA control
as well to provide ALSA clients both interfaces.

Signed-off-by: Dylan Van Assche <me@dylanvanassche.be>