configs: sdm845: enable LED_TRIGGER_PATTERN as built-in
LED_TRIGGER_PATTERN is necessary for feedbackd to function correctly.
However, it doesn't get enabled by default when build as module,
build as built-in to avoid this problem so the notification LED
blinks correctly.

This works around the problem that LED_TRIGGER_PATTERN is not probed
automatically when the Qualcomm LPG driver is used for SDM845 devices.