FIXUP: drm: panel: lg-sw43408: allow turning off and cleanup
Whatever init sequence we have for this panel isn't capable of
initialising it completely, toggling the reset gpio ever causes the
panel to die, add comments explaining this and clean up the driver.

This also flips the orientation of the reset GPIO to make more logical
sense, driving it high now means in reset, low means out of reset.