configs: sdm845: always mark BT_BNEP as module
CONFIG_BT_BNEP should always be build as a module otherwise it may never
be initialized at boot. Without this module, BlueZ will not start up its
networking plugin to provide Bluetooth PAN support.

BNEP support as builtin may work if the Bluetooth driver is builtin
(which is not the case here), but I haven't verified that.