configs: sdm845: always mark BT_RFCOMM as module
CONFIG_BT_RFCOMM should always be build as a module otherwise it may never
be initialized at boot. Without this module, BlueZ will not enable
profiles depending on RFCOMM such as HFP.

RFCOMM support as builtin may work if the Bluetooth driver is builtin
(which is not the case here), but I haven't verified that.