HACK: ASoC: codecs: wcd934x: adjust the max volume for RX digital volume and decimator volume
The Earpiece, headphones and mics appears to be having low volume for some reason.
Could also be because of some recent changes in pulseaudio?
Adjusting the maximum dB range of these digital volume helps with temporarily fixing this issue.
Downstream mixer paths xml set 84 as default volumes for these in many devices.
So I assume that is maximum volume?
However 84dB was a bit loud and since users can increase the volume through userspace workarounds to beyond 100%,
75dB is chosen as safe/optimal maximum value for now, so that we don't damage the devices by chance.
This can adjusted later on if needed.