net: ipa: move wakeup event handling out of ipa_suspend_handler()
Move the flag manipulation and pm_wakeup notification to it's own
function: ipa_wakeup_triggered() and make it return a boolean value,
true if we are currently in system suspend and IPA_POWER_FLAG_RESUMED
was unset.

This function will be used in the ipa IRQ handler to determine if the
in-progress IRQ was the cause of a system resume. For now retain the
original functionality by calling it from ipa_suspend_handler().

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>