net: ipa: migrate to dev_pm based wakeup handling
Replace the call to enable_irq_wake() with one to the
dev_pm_set_wake_irq() wrapper. This will associate the IRQ as a wakeirq
in the dev_pm framework, allowing the it to (mostly) manage it for us.

Most significantly, this allows userspace to disable the wakeup source
associated with the IPA, and consequently prevent the IPA from causing a
wakeup.

The IPA suspend and resume callbacks and interrupt handler are adjusted
to ensure that interrupts are properly sequenced with systemd
suspend/resume events. For the case where wakeup is enabled we defer
interrupt processing until the ipa_resume() callback, when wakeup is
disabled we ensure that the IRQ is masked until resuming. In both cases
this is done to ensure that the interrupt handler doesn't run whilst PM
runtime is disabled for the device.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>