arm64: dts: qcom: sdm845-shift-axolotl: EUD clk and regulator
downstream for 845 has the EUD vote on the USB_PHY_CFG_AHB2PHY clock,
something about a NOC error if we try and talk to the EUD while the
clock is off.

Also mark ldo24 as always on, downstream says "eud does not retain
interrupt mask when ldo24 is turned off" and works around it by setting
up a notifier for when the regulator is enabled and writes the interrupt
mask then. bleh

https://github.com/LineageOS/android_kernel_oneplus_sdm845/commit/5a4977819b83b3b70169397040da8a3508f27cab