NOTE: this is probably all for moot :(
The SHIFT6mq has QFPROM OEM Config Row 0 MSB set to 0x00000004, setting
just bit 2, which is EUD_DISABLE.

On both the 6mq and OnePlus 6, the EUD can be enabled and seems to
sorta-work, my PC hates the USB descriptors but on the Yoga C630 laptop
it behaves a bit better, i get two devices with idProduct 0x9500 and
0x9501 respectively. 0x9501 is the EUD control interface.

I suspect that this is enough for the host software to go "ah, sorry no
can do", i imagine that if it wasn't disabled we would have a bunch of
other devices as well. But I can't be sure.