net: ipa: use _noirq suspend/resume pm_ops
The next patch will associate the IPA irq as a wakeirq and enable the
wakeirq handling in the PM core code. Because the interrupt handler
requires pm_runtime to be enabled, lets delay our calls to
pm_runtime_force_suspend() until we know the IRQ handler won't be
running, and similarly do pm_runtime_force_resume() earlier.

This ensures that we don't end up racing with the handler if it fires
during the suspend process. And that if there are pending interrupts
when resuming, pm_runtime will be enabled before the handler runs.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>