net: ipa: associate the ipa IRQ as a wakeirq
This allows the dev PM core to manage the interrupt for us by
automatically arming and disarming it, as well as waking up the system
when it fires and then deferring it until IRQs are enabled after the
system resume_noirq() phase.

This ensures that we will never call the IRQ handler without pm_runtime
enabled, and that if wakeup is disabled for the IPA then the IRQ will be
correctly postponed until the system resumes.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>