net: ipa: keep ipa interrupt disabled in suspend
The IPA interrupt is capable of waking the system up. However it
requires pm_runtime to be enabled for the handler to be able to read
registers and determine the cause of the interrupt. The next patch will
enable wakeirq functionality offered by the PM core, this would lead to
concurrency issues as it keeps pm_runtime disabled until after calling
driver resume_early() callbacks, but enables interrupts after calling
resume_noirq(). This results in a short period of time where the IRQ
handler may run with pm_runtime disabled.

By disabling the IRQ during ->suspend and re-enabling it during ->resume
we avoid this concurrency issue and allow the IRQ to be processed only
once pm_runtime() is available again.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>