ci: add gitlab ci
Gitlab CI with the following features:
 * Build validation (with pmbootstrap)
 * automated boot-test (WIP, need to test running from an MR with manual
   review
 * stable merging
 * automated package upgrades in pmaports

Based on the branch naming scheme, either the release or dev pipeline
will be triggered. In -dev branches the kernel will be compiled and
boot-tested using cdba.

In -release branches, the pipeline will merge the latest stable patches
in, then build, boot test, and finally open an MR to pmaports to upgrade
the kernel once all tests pass.

This is a fairly rough first pass, and it definitely needs a lot of
improvement.