drm/panel: add support for the SHIFT6mq variant of rm69299
The SHIFT6mq ships with a Visionox RM69299 panel, however it uses a
different resolution and vastly different init sequence. Without access
to proprietary documentation or the version of the panel supported by
the existing driver it is more complicated to support both variants in a
single driver. Instead, let's introduce a new driver for the SHIFT6mq
variant of the panel.

Generated with the Linux MDSS DSI panel driver generator [1] with manual
additions for backlight and regulators.

[1]: https://github.com/msm8916-mainline/linux-mdss-dsi-panel-driver-generator

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>