Input: TCM oncell code dump
need to figure out the authorship and such. This is a total rewrite
using Frieder's cut-down TCM oncell driver as inspiration, but largely
written iteratively by dumping all the reads and reports from the touch
IC and just doing the bare minimum to get this functional.

The configurable report_config stuff is skipped as it seems the S3908
(at least the one in the op8t) doesn't support writing the
configuration. Therefore we can just use a packed struct to represent
the report data.