drm/panel: sofef00: proper power sequencing
There are two additional regulators on this panel, vci and poc. Use the
regulator bulk API to fetch and enable all regulators at once.

In addition, let's also fix how we model the power sequencing of this
panel. On the devices it's used on, the panel is powered up and
configuring by the bootloader during boot to display a splash image.
This makes the first .prepare() call totally redundant, as all the
supplies are on and the panel has been programmed correctly.

The hygenic thing to do would be to fully power sequence the panel to be
super sure that it's initialised the way we expect. However, the only
user of this panel (the OnePlus 6) isn't likely to get any breaking
changes to the bootloader. So let's just skip the first prepare() if the
regulators are already enabled.

This gives us a fairly smooth transition, with only very minor visual
artifacting (a while horizontal line) when the DSI engine starts up.

For future reference, if we replace the "return 0" with a call to
sofef00_panel_off() followed by regulator_bulk_disable() and a sleep, we
can then continue the prepare as normal and the panel behaves just the
same.

Once the panel is powered off for the first time, we will then do a full
power up during prepare.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>